LayoutTests/css3/flexbox/bug633212.html

<!DOCTYPE html>
<style>
html, body {
  margin: 0;
}

#flexbox {
    display: flex;
    flex-direction: column;
}

#flexitem {
    outline: 1px solid black;
    width: 100%;
    height: 166px;
    position: relative;
    background: green;
}

#abspos {
    position: absolute;
}

#container {
    background: red;
}
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>
function runTest() {
  test(function() {
    assert_equals(document.getElementById('flexbox').offsetWidth,
      document.getElementById('flexitem').offsetWidth);
  }, "sizes");
}
</script>

<body onload="runTest()">
<div id="container">
  <div id="flexbox">
      <div id="flexitem">
          <div id="abspos">
              <svg></svg>
          </div>
      </div>
  </div>
</div>
